--- src-tauri/.cargo/config.toml
@@ -1,7 +1,0 @@
-[target.x86_64-pc-windows-msvc]
-rustflags = ["-Ctarget-feature=+crt-static", "-Clink-arg=-fuse-ld=lld"]
-[target.aarch64-apple-darwin]
-rustflags = ["-Ctarget-feature=+crt-static"]
-
-[env]
-ORT_LIB_LOCATION = "./lib"
