+++ src-tauri/Cargo.toml
@@ -1,9 +1,15 @@
 [workspace]
 members = [
-    "src-crates/app-services",
-    "src-crates/app-os",
     "src-crates/app-shared",
     "src-crates/app-utils",
+    "src-crates/app-services",
+    "src-crates/app-scroll-screenshot-service",
+    "src-crates/app-os",
+    "src-crates/tauri-commands/core",
+    "src-crates/tauri-commands/file",
+    "src-crates/tauri-commands/ocr",
+    "src-crates/tauri-commands/screenshot",
+    "src-crates/tauri-commands/scroll-screenshot",
 ]
 
 [package]
@@ -36,7 +42,41 @@ image = { version = "^0.25", features = ["png", "webp"] }
 imageproc = "^0.25"
 enigo = "^0.5"
 tokio = "^1.45"
-tauri = { version = "^2", features = [
+tauri = { version = "^2" }
+rayon = "^1.10"
+regex = "^1.11"
+zune-jpegxl = "^0.4"
+zune-core = "^0.4"
+
+# 项目内依赖
+snow-shot-app-services = { path = "src-crates/app-services" }
+snow-shot-app-os = { path = "src-crates/app-os" }
+snow-shot-app-utils = { path = "src-crates/app-utils" }
+snow-shot-app-shared = { path = "src-crates/app-shared" }
+snow-shot-app-scroll-screenshot-service = { path = "src-crates/app-scroll-screenshot-service" }
+snow-shot-tauri-commands-core = { path = "src-crates/tauri-commands/core" }
+snow-shot-tauri-commands-file = { path = "src-crates/tauri-commands/file" }
+snow-shot-tauri-commands-ocr = { path = "src-crates/tauri-commands/ocr" }
+snow-shot-tauri-commands-screenshot = { path = "src-crates/tauri-commands/screenshot" }
+snow-shot-tauri-commands-scroll-screenshot = { path = "src-crates/tauri-commands/scroll-screenshot" }
+
+[dependencies]
+snow-shot-app-shared = { workspace = true }
+snow-shot-app-services = { workspace = true }
+snow-shot-app-os = { workspace = true }
+snow-shot-app-scroll-screenshot-service = { workspace = true }
+snow-shot-tauri-commands-core = { workspace = true }
+snow-shot-tauri-commands-file = { workspace = true }
+snow-shot-tauri-commands-ocr = { workspace = true }
+snow-shot-tauri-commands-screenshot = { workspace = true }
+snow-shot-tauri-commands-scroll-screenshot = { workspace = true }
+
+# workspace 依赖
+tokio = { workspace = true }
+log = { workspace = true }
+
+# tauri 相关
+tauri = { workspace = true, features = [
     "protocol-asset",
     "macos-private-api",
     "tray-icon",
@@ -44,34 +84,6 @@ tauri = { version = "^2", features = [
     "image-ico",
     "image-png",
 ] }
-rayon = "^1.10"
-regex = "^1.11"
-zune-jpegxl = "^0.4"
-zune-core = "^0.4"
-
-[dependencies]
-# 项目内依赖
-snow-shot-app-services = { path = "src-crates/app-services" }
-snow-shot-app-os = { path = "src-crates/app-os" }
-snow-shot-app-utils = { path = "src-crates/app-utils" }
-
-# workspace 依赖
-serde_json = { workspace = true }
-serde = { workspace = true }
-xcap = { workspace = true }
-device_query = { workspace = true }
-log = { workspace = true }
-image = { workspace = true }
-imageproc = { workspace = true }
-enigo = { workspace = true }
-tokio = { workspace = true }
-rayon = { workspace = true }
-regex = { workspace = true }
-zune-jpegxl = { workspace = true }
-zune-core = { workspace = true }
-
-# tauri 相关
-tauri = { workspace = true }
 tauri-plugin-log = "^2"
 tauri-plugin-fs = "^2"
 tauri-plugin-dialog = "^2"
@@ -83,16 +95,6 @@ tauri-plugin-clipboard = "^2"
 tauri-plugin-process = "^2"
 tauri-plugin-notification = "^2"
 tauri-plugin-os = "^2"
-
-# 其他依赖
-paddle-ocr-rs = "0.5.2"
-# paddle-ocr-rs = { git = "https://github.com/mg-chao/paddle-ocr-rs.git", branch = "main" }
-num_cpus = "^1.16"
-
-base64 = "^0.22"
-bytemuck = "^1.23"
-get-selected-text = "^0.1"
-clipboard-rs = "^0.3"
 
 [target.'cfg(any(target_os = "windows"))'.dependencies]
 windows = { version = "0.61.1", features = [
@@ -121,5 +123,3 @@ tauri-plugin-autostart = "2"
 tauri-plugin-global-shortcut = "^2"
 tauri-plugin-single-instance = "2"
 
-[profile.dev]
-opt-level = 1
