+++ src-tauri/src-crates/app-scroll-screenshot-service/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "snow-shot-app-scroll-screenshot-service"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde = { workspace = true }
+image = { workspace = true }
+imageproc = { workspace = true }
+rayon = { workspace = true }
+
+fast_image_resize = { version = "^5.2", features = ["rayon"] }
+hora = { version = "^0.1.1" }
