+++ src-tauri/src-crates/app-scroll-screenshot-service/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod scroll_screenshot_image_service;
+pub mod scroll_screenshot_service;
