+++ src-tauri/src-crates/app-scroll-screenshot-service/src/scroll_screenshot_service.rs
