+++ src-tauri/src-crates/app-services/src/lib.rs
@@ -1,4 +1,2 @@
 pub mod free_drag_window_service;
-pub mod scroll_screenshot_image_service;
-pub mod scroll_screenshot_service;
 pub mod video_record_service;
