+++ src-tauri/src-crates/app-shared/Cargo.toml
@@ -4,3 +4,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+enigo = { workspace = true }
