+++ src-tauri/src-crates/app-utils/Cargo.toml
@@ -10,3 +10,4 @@ zune-jpegxl = { workspace = true }
 zune-core = { workspace = true }
 device_query = { workspace = true }
 image = { workspace = true }
+log = { workspace = true }
