+++ src-tauri/src-crates/tauri-commands/file/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "snow-shot-tauri-commands-file"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+image = { workspace = true }
+zune-core = { workspace = true }
+zune-jpegxl = { workspace = true }
+tauri = { workspace = true }
+
+base64 = "^0.22"
