+++ src-tauri/src-crates/tauri-commands/ocr/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "snow-shot-tauri-commands-ocr"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tauri = { workspace = true }
+tokio = { workspace = true }
+serde = { workspace = true }
+image = { workspace = true }
+
+paddle-ocr-rs = "0.5.2"
+# paddle-ocr-rs = { git = "https://github.com/mg-chao/paddle-ocr-rs.git", branch = "main" }
+num_cpus = "^1.16"
