+++ src-tauri/src-crates/tauri-commands/screenshot/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "snow-shot-tauri-commands-screenshot"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+image = { workspace = true }
+tauri = { workspace = true }
+tokio = { workspace = true }
+xcap = { workspace = true }
+serde = { workspace = true }
+log = { workspace = true }
+
+snow-shot-app-utils = { workspace = true }
+snow-shot-app-os = { workspace = true }
