+++ src-tauri/src-crates/tauri-commands/scroll-screenshot/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "snow-shot-tauri-commands-scroll-screenshot"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+image = { workspace = true }
+serde = { workspace = true }
+tauri = { workspace = true }
+tokio = { workspace = true }
+xcap = { workspace = true }
+
+snow-shot-app-utils = { workspace = true }
+snow-shot-app-scroll-screenshot-service = { workspace = true }
